test(measure): share the object-panel DOM stub across its three suites
The stub was declared independently in objectMetricsUnitNormalisation,
objectPanelSpace and objectPanelExport, which Sonar counted as duplicated
lines on every new object-panel test. tests/support/objectPanelDom.ts now
holds the superset, including the recursive textContent those suites need
and findByText.
